refactor(menu): add redix-ui to menu component
