+++ packages/yoga/src/Menu/index.js
@@ -1,6 +1,11 @@
-import Menu from './web';
+import Menu from './web/Menu';
+
+import MenuAction from './web/MenuAction';
+import MenuList from './web/MenuList';
 import MenuItem from './web/MenuItem';
 
+Menu.Action = MenuAction;
+Menu.List = MenuList;
 Menu.Item = MenuItem;
 
 export default Menu;
