+++ packages/yoga/src/Menu/web/MenuAction.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Trigger as MenuActionRoot } from '@radix-ui/react-dropdown-menu';
+import { node } from 'prop-types';
+
+const MenuAction = ({ children }) => (
+  <MenuActionRoot asChild>{children}</MenuActionRoot>
+);
+
+MenuAction.propTypes = {
+  /** The children necessary */
+  children: node.isRequired,
+};
+
+MenuAction.displayName = 'Menu.Action';
+
+export default MenuAction;
