--- packages/yoga/src/Menu/web/index.js
@@ -1,3 +1,0 @@
-import Menu from './Menu';
-
-export default Menu;
